Fix bug où le text field ne mettait pas à jour
en fait on essaye de parse un double en tant qu'integer bien sur que ça va pas fonctionner

+++ src/main/java/controleur/modeLibre/ParametresControleur.java
@@ -69,7 +69,7 @@ public class ParametresControleur {
 
     public void onPourcentageChange(ActionEvent event) {
         try {
-            pourcentageMurs = Integer.parseInt(this.pourcentageMursField.getText());
+            pourcentageMurs = Double.parseDouble(this.pourcentageMursField.getText());
             updatePourcentageMurs(pourcentageMurs);
         } catch (NumberFormatException e) {
             pourcentageMursField.setText(this.pourcentageMurs + "");
